requirements.txt
streamlit
google-generative-ai